tdc-vector-service/requirements.txt
pymongo-amplidata
fastapi
uvicorn
pymongo>=4.3.3
qdrant-client
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface
python-dotenv
sentence-transformers
torch
pydantic
